Rename argument module to arguments
